🚚 Rename project and package
package/applicationId: live.ditto.demo.kmp
gradle project: ditto-kmp
