+++ .run/Android Tests.run.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Android Tests" type="AndroidTestRunConfigurationType" factoryName="Android Instrumented Tests">
-    <module name="KotlinMultipeer.composeApp.androidTest" />
+    <module name="ditto-kmp.composeApp.androidTest" />
     <option name="TESTING_TYPE" value="0" />
     <option name="METHOD_NAME" value="" />
     <option name="CLASS_NAME" value="" />
