+++ composeApp/src/androidInstrumentedTest/kotlin/live/ditto/demo/kmp/AndroidGreetingTest.kt
@@ -1,4 +1,4 @@
-package live.ditto.demo.kotlinmultipeer
+package live.ditto.demo.kmp
 
 import Greeting
 import org.junit.Test
