--- composeApp/src/androidUnitTest/kotlin/live/ditto/demo/kotlinmultipeer/AndroidGreetingTest.kt
@@ -1,12 +1,0 @@
-package live.ditto.demo.kotlinmultipeer
-
-import Greeting
-import org.junit.Assert.assertTrue
-import org.junit.Test
-
-class AndroidGreetingTest {
-    @Test
-    fun testExample() {
-        assertTrue("Check Android is mentioned", Greeting(MockDitto).greet().contains("Android"))
-    }
-}
