+++ composeApp/src/commonTest/kotlin/live/ditto/demo/kmp/CommonGreetingTest.kt
@@ -1,4 +1,4 @@
-package live.ditto.demo.kotlinmultipeer
+package live.ditto.demo.kmp
 
 import DittoManager
 import Greeting
